Restructure code for customer insights
