+++ src/pages/CustomerInsights/CustomerInsights.jsx
@@ -11,256 +11,7 @@ const BUSINESS_OWNERS = "Business Owners"
 import { LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip, Legend } from 'recharts';
 import { Radar, RadarChart, PolarGrid, PolarAngleAxis, PolarRadiusAxis, ResponsiveContainer } from 'recharts';
 import { BarChart, Bar, Rectangle } from 'recharts';
-
-export const data_5 = [
-    {
-        name: '1 Nov',
-        signups: 60,
-        amt: 2400,
-    },
-    {
-        name: '2 Nov',
-        signups: 55,
-        amt: 2210,
-    },
-    {
-        name: '3 Nov',
-        signups: 70,
-        amt: 2290,
-    },
-    {
-        name: '4 Nov',
-        signups: 81,
-        amt: 2000,
-    },
-    {
-        name: '5 Nov',
-        signups: 65,
-        amt: 2181,
-    },
-    {
-        name: '6 Nov',
-        signups: 77,
-        amt: 2500,
-    },
-    {
-        name: '7 Nov',
-        signups: 78,
-        amt: 2100,
-    },
-    {
-        name: '8 Nov',
-        signups: 79,
-        amt: 2100,
-    },
-    {
-        name: '9 Nov',
-        signups: 84,
-        amt: 230,
-    },
-    {
-        name: '10 Nov',
-        signups: 90,
-        amt: 234,
-    },
-    {
-        name: '11 Nov',
-        signups: 92,
-        amt: 2100,
-    },
-    {
-        name: '12 Nov',
-        signups: 93,
-        amt: 2100,
-    },
-    {
-        name: '13 Nov',
-        signups: 94,
-        amt: 400,
-    },
-    {
-        name: '14 Nov',
-        signups: 95,
-        amt: 300,
-    },
-    {
-        name: '15 Nov',
-        signups: 97,
-        amt: 770,
-    },
-    {
-        name: '16 Nov',
-        signups: 95,
-        amt: 238,
-    },
-    {
-        name: '17 Nov',
-        signups: 96,
-        amt: 700,
-    },
-];
-
-const data_3 = [
-    {
-        name: 'Jan 2023',
-        signups: 100,
-    },
-    {
-        name: 'Feb 2023',
-        signups: 112,
-    },
-    {
-        name: 'Mar 2023',
-        signups: 134,
-    },
-    {
-        name: 'Apr 2023',
-        signups: 99,
-    },
-    {
-        name: 'May 2023',
-        signups: 345,
-    },
-    {
-        name: 'Jun 2023',
-        signups: 233,
-    },
-    {
-        name: 'Jul 2023',
-        signups: 210,
-    },
-    {
-        name: 'Aug 2023',
-        signups: 180,
-    },
-    {
-        name: 'Sept 2023',
-        signups: 188,
-    },
-    {
-        name: 'Oct 2023',
-        signups: 134,
-    },
-    {
-        name: 'Nov 2023',
-        signups: 566,
-    },
-];
-
-
-const data_4 = [
-    {
-        name: 'Jan 2023',
-        churn: 20,
-    },
-    {
-        name: 'Feb 2023',
-        churn: 23,
-    },
-    {
-        name: 'Mar 2023',
-        churn: 24,
-    },
-    {
-        name: 'Apr 2023',
-        churn: 22,
-    },
-    {
-        name: 'May 2023',
-        churn: 21,
-    },
-    {
-        name: 'Jun 2023',
-        churn: 15,
-    },
-    {
-        name: 'Jul 2023',
-        churn: 10,
-    },
-    {
-        name: 'Aug 2023',
-        churn: 8,
-    },
-    {
-        name: 'Sept 2023',
-        churn: 12,
-    },
-    {
-        name: 'Oct 2023',
-        churn: 9,
-    },
-    {
-        name: 'Nov 2023',
-        churn: 5,
-    },
-];
-
-
-const data_1 = [
-    {
-        subject: 'Invest Program',
-        A: 120,
-        fullMark: 150,
-    },
-    {
-        subject: 'Earn and Learn',
-        A: 98,
-        fullMark: 150,
-    },
-    {
-        subject: 'Online Banking',
-        A: 86,
-        fullMark: 150,
-    },
-    {
-        subject: 'Loans',
-        A: 99,
-        fullMark: 150,
-    },
-    {
-        subject: 'Bank Account',
-        A: 85,
-        fullMark: 150,
-    },
-    {
-        subject: 'Bank Card',
-        A: 65,
-        fullMark: 150,
-    },
-];
-
-const data_2 = [
-    {
-        subject: 'Robo Advisor',
-        A: 120,
-        fullMark: 150,
-    },
-    {
-        subject: 'Shopback',
-        A: 98,
-        fullMark: 150,
-    },
-    {
-        subject: 'Stashaway',
-        A: 86,
-        fullMark: 150,
-    },
-    {
-        subject: 'GoodBudget',
-        A: 99,
-        fullMark: 150,
-    },
-    {
-        subject: 'PocketGuard',
-        A: 85,
-        fullMark: 150,
-    },
-    {
-        subject: 'Mobills',
-        A: 65,
-        fullMark: 150,
-    },
-];
+import { data_1, data_2, data_3, data_4, data_5 } from './CustomerInsights.config';
 
 const CustomerInsights = () => {
 
@@ -414,8 +165,6 @@ const CustomerInsights = () => {
                     </div>
                 </div>
             </div>
-
-
         </Layout>
     );
 };
